gdb: introduce limited array lengths while printing values
This commit introduces the idea of loading only part of an array in
order to print it, what I call "limited length" arrays.

The motivation behind this work is to make it possible to print slices
of very large arrays, where very large means bigger than
max-value-size.

Consider this GDB session with the current GDB:

  (gdb) set max-value-size 100
  (gdb) p large_1d_array
  value requires 400 bytes, which is more than max-value-size
  (gdb) p -elements 10 -- large_1d_array
  value requires 400 bytes, which is more than max-value-size

notice that the request to print 10 elements still fails, even though
10 elements should be less than the max-value-size. With a patched
version of GDB:

  (gdb) p -elements 10 -- large_1d_array
  $1 = {0, 1, 2, 3, 4, 5, 6, 7, 8, 9...}

So now the print has succeeded.  The only remaining problem is that
the '$1 = ' implies that the value of the array has been placed into
the value history, so what happens if the tries to reprint the full
array without the print elements?  My solution for now is:

  (gdb) p $1
  $2 = <unavailable>

This patch is currently RFC, I would like to hear what people think
about both the idea in general, and the approach taken.

One question I have is whether the value history problem would be
better addressed in a different way, for example, we could just drop
the '$1 = ' for values that are not being added into the history, so
things would look like:

  (gdb) p -elements 10 -- large_1d_array
  {0, 1, 2, 3, 4, 5, 6, 7, 8, 9...}

which might be a better solution.  Another possibility would be to tag
the "unavailable" value with a reason field so we could do something
like:

  (gdb) p $1
  $2 = <unavailable: original value was too large>

which is slightly more informative, but clearly is a more invasive
change to the value structure.

The current solution relies on loading part of the array value into
GDB's memory, this is done because very early in the value printing
code (in value_check_printable) we call value_entirely_optimized_out,
which loads the values complete contents into memory.

But I think Tom was looking into having the value optimized out checks
not actually load the value in at one point, so maybe, if that change
landed, then we could investigate the possibility of having the array
printing code only load the elements from the target one at a
time (with the dcache providing some optimisation), which might avoid
the need to perform the current partial load?

Anyway, I'd be interested to hear people's thoughts; is this a
valuable change?  Which approach seems like the right way to go?