[wip] gdb/python: fix failure caused by searching for a user reg
The problem I attempt to fix in this issue is that, if a user searches
for a user-reg then we will return a gdb.RegisterDescriptor, but the
descriptor could have no name, and thus, could cause a failure if we
try to print the register.

For example, this will trigger the issue on RISC-V GDB:

  arch = gdb.selected_inferior().architecture()
  regs_iter = arch.registers()
  r = regs_iter.find('x0')
  print("Got register: %s" % r)
  for r in regs_iter:
      print("Reg: %s" % r)

I wonder if now is a good time to improve how we use user-regs?
Mostly, these are used for register aliases, so I wonder if we
could/should move more to an official position that these are just
that, register aliases, then, when we lookup a register number using
an alias, we could/should get back the "official" number, not the
number of the user-reg which is the alias.

This would require an audit of all uses of user-regs to check that
they really are just used for aliases.