gdbsupport: better detection of -Wmissing-prototypes support
When building using ccache I notice lots of warnings like this while
building GDB:

  cc1plus: warning: command line option ‘-Wmissing-prototypes’ is valid for C/ObjC but not for C++

This is a little strange as the configure macro, AM_GDB_WARNINGS,
should figure out which warning flags are valid, and which are not.

It turns out that there's an issue with ccache and gcc relating to
option ordering, these are described here:

  https://github.com/ccache/ccache/issues/738
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=87161

Basically, ccache reorders some command line options, so this:

  -Werror -Wmissing-prototypes

becomes:

  -Wmissing-prototypes -Werror

And, so, when g++ sees the -Wmissing-prototypes option, it has not yet
seen -Werror, and as -Wmissing-prototypes is not supported in g++, a
warning is immediately issued.

What this means is that when the AM_GDB_WARNINGS macro tries to test
if -Wmissing-prototypes is support on g++ (and if the user is using
ccache), then all the macro ever sees is a warning, not an error, and
so the macro assumes that -Wmissing-prototypes is supported.

Now the AM_GDB_WARNINGS makes use of AM_COMPILE_IFELSE to compile a
test file to object format.  IF instead we use AM_LINK_IFELSE then the
compiler will be invoked to take the source file all the way to
executable format.

Now, when ccache sees the compiler invoked to do a full build (to
executable), ccache does not activate, and there is not command line
flag reordering, and so, we get an error for -Wmissing-prototypes
rather than a warning, and this flag will not be used (for g++).