gdb/gdbserver: fix uninstall makefile rules
I noticed that after running:

  $ make install-gdb install-gdbserver

and then:

  $ make -C gdb uninstall && make -C gdbserver uninstall

the install destination looks like:

  install
  |-- bin
  |   `-- gdb-add-index
  |-- include
  |   `-- gdb
  |       `-- jit-reader.h
  |-- lib
  |   `-- libinproctrace.so
  `-- share
      |-- gdb
      |-- info
      |   `-- dir
      `-- man
          |-- man1
          `-- man5

We've left 3 files behind.  This is because the uninstall rules in
gdb/ and gdbserver/ have gotten out of sync with the install rules.
This commit adds the missing commands to the uninstall rules to delete
the extra files.