gdb/gdbserver: add a '--no-escape-args' command line option
This introduces a new '--no-escape-args' option for gdb and gdbserver.

I (Andrew Burgess) have taken this patch from this series:

    https://inbox.sourceware.org/gdb-patches/20211022071933.3478427-1-m.weghorn@posteo.de/

I wasn't completely convinced by the solution proposed in that series,
so I wanted to offer my own ideas.  However, this change I think is
still an important one, and as the code here is pretty much identical
to the original suggestion, I've left the patch author unchanged.  Any
bugs introduced by my modifications to the original patch should be
considered mine.  The original commit also lacked documentation and
tests, which I (Andrew) have added.

When the startup-with-shell option is enabled, args passed directly on
'gdb --args <args>' or 'gdbserver <args>' invocation are by default
escaped so that they are passed to the inferior as passed on
invocation, no globbing or variable substitution happens when the args
are passed to the shell that is started by the GDB process later.

For gdbserver, this is the case since commit:

  commit bea571ebd78ee29cb94adf648fbcda1e109e1be6
  Date:   Mon May 25 11:39:43 2020 -0400

      Use construct_inferior_arguments which handles special chars

Only arguments set via 'set args <args>', 'run <args>', or through the
Python API are not escaped.

For the 'gdb --args' case, directly settings unescaped args on gdb
invocation is possible e.g. by using the "--eval-command='set args
<args>'", while this possibility does not exist for gdbserver.

The new '--no-escape-args' options disable the escaping of args passed
directly on invocation for both GDB and gdbserver.

For gdbserver, using this new option allows having the behaviour from
before commit bea571ebd78ee29cb94adf648fbcda1e109e1be6, while keeping
the default behaviour unified between GDB and GDBserver.

For GDB the --no-escape-args option can be used as a replacement for
--args, like this:

  shell> gdb --no-escape-args my-program arg1 arg2 arg3

While for gdbserver, the --no-escape-args option is a flag, which can
be used like:

  shell> gdbserver --no-escape-args --once localhost:54321 \
             my-program arg1 arg2 arg3

Co-Authored-By: Andrew Burgess <aburgess@redhat.com>
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=28392