gdb: Support some escaping of args with startup-with-shell being off
Previously, 'execv_argv::init_for_no_shell', which is called
from the 'fork_inferior' overload that gets all args in a single string
had a custom implementation to split the string containing all args
into an array of single args when the 'startup-with-shell' option is
off (which is on by default).

That implementation was just splitting at whitespace characters and
ignoring any quoting, so e.g. args set for target-board unix using

    set startup-with-shell off
    set args "first arg" second_arg

would result in three args {'"first', 'arg"', 'second_arg'} being
passed to the inferior.

Replace this custom splitting of the args string into single args
and use the existing 'buildargv' function instead, which supports
some basic quoting and escaping, so e.g. the above results in two
args {'first arg', 'second_arg'} being passed to the debuggee,
as is the case when the 'startup-with-shell' option is on.
The same splitting of args was already used for target
extended-remote before (s. args splitting done in
'remote_target::extended_remote_run', and the 'fork_inferior'
overload taking args as a vector is used later on, so no additional
splitting is done then).
Thus, using the same here makes the behaviour more consistent
between GDB and GDBserver as well.

This changes the 'execv_argv' class accordingly and drops the
optimization to have all the 'char *' in 'm_argv' point to a single
string rather than allocating a separate string for each arg.

With this in place, quoting or escaping whitespace characters
is now supported with startup-with-shell being off, and literal
quotes can be inserted by escaping them, so e.g. the previous
behaviour of the above

    set startup-with-shell off
    set args "first arg" second_arg

example can now be reached by using either

    set startup-with-shell off
    set args \"first arg\" second_arg

or

    set startup-with-shell off
    set args '"first' 'arg"' second_arg

Also extend the existing 'gdb.base/inferior-args.exp' test
to cover the case where 'startup-with-shell' is off.