gdb: Support some escaping of args with startup-with-shell being off
I (Andrew Burgess) have taken this patch from this series:

  https://inbox.sourceware.org/gdb-patches/20211022071933.3478427-1-m.weghorn@posteo.de/

I started off reviewing that series, but wanted to explore some
alternative solutions the problem of passing inferior arguments to the
remote target.  However, this patch I think is super useful, so I've
taken it most as it was in the original series.

I have made a few minor cleanups, and I've also added some more
tests.  Any bugs should be considered mine (Andrew), but I've left the
original author in place as the patch is mostly their work.

Previously execv_argv::init_for_no_shell (gdb/nat/fork-inferior.c),
which is called from the execv_argv constructor had a custom
implementation to split the string containing all args into an array
of single args when the 'startup-with-shell' option is off (which is
on by default).

That implementation was just splitting at whitespace characters and
ignoring any quoting, so for example this:

    set startup-with-shell off
    set args "first arg" second_arg

would result in three arguments '"first', 'arg"', and 'second_arg'
being passed to the inferior.

Replace this custom arg splitting with a use of the existing gdb_argv
class (which uses the libiberty buildargv function).  This does a
better job of supporting quoting and escaping, so for the example
given above we how pass two args {'first arg', 'second_arg'} to the
inferior, as is the case when the 'startup-with-shell' option is on.

This commit changes the 'execv_argv' class accordingly and drops the
optimization to have all the 'char *' in 'm_argv' point to a single
string rather than allocating a separate string for each arg.

With this in place, quoting or escaping whitespace characters
is now supported with startup-with-shell being off, and literal
quotes can be inserted by escaping them, so e.g. the previous
behaviour of the above

    set startup-with-shell off
    set args "first arg" second_arg

example can now be reached by using either

    set startup-with-shell off
    set args \"first arg\" second_arg

or

    set startup-with-shell off
    set args '"first' 'arg"' second_arg

This commit also extends the three tests that cover inferior argument
passing to cover the case where 'startup-with-shell' is off.  All of
these new tests pass for native targets, but there are still problems
when using remote targets.  For now I have (rather crudely) disabled
these tests for remote targets.  Don't worry too much, the iffy
testsuite code will be removed by the end of this series, by which
point remote targets will be as functional as native targets (for
argument passing at least).

Co-Authored-By: Andrew Burgess <aburgess@redhat.com>
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=28392