gdb/python: make more use of host_string_to_python_string
We have a function host_string_to_python_string, which is a wrapper
around PyString_Decode, which, on Python 3, is an alias for
PyUnicode_Decode.

However, there are a few places where we still call PyUnicode_Decode
directly.

This commit replaces all uses of PyUnicode_Decode with calls to
host_string_to_python_string instead.

To make the use of host_string_to_python_string easier, I've added a
couple of overloads for this function in python-internal.h, these all
just forward their calls onto the single base implementation.  The
signatures of all host_string_to_python_string overloads are:

  gdbpy_ref<> host_string_to_python_string (const char *str, size_t len);
  gdbpy_ref<> host_string_to_python_string (const char *str)
  gdbpy_ref<> host_string_to_python_string (const string_file &str)

For Python 3 PyString_Decode is setup (in python-internal.h) as an
alias for PyUnicode_Decode, so there should certainly be no user
visible changes in this case.

For Python 2 this commit will change the behaviour.  Previously by
calling PyUnicode_Decode we would have been returning a Unicode
object.  Now, after calling host_string_to_python_string, we will have
a str object.

I've checked the GDB documentation, and, as far as I can tell, the
methods I've touched were all documented as returning a string, or in
the gdb.Command case, take a string as an argument, and my
understanding is that for Python 2, string generally means str.  So I
think the new behaviour would be more expected.

A different solution, that would also make things more consistent in
the Python 2 world, would be to have host_string_to_python_string
always return a Unicode object.  However, I've been reading this page:

  https://pythonhosted.org/kitchen/unicode-frustrations.html

item #3 recommends that unicode strings should be converted to str
objects before being printed (in Python 2).  That would mean that
users should have been adding .encode() calls to the output of the
routines I've changed in this commit (if they wanted to print that
output), which is not something I think is made clear from the GDB
documentation.