gdb/python: make more use of host_string_to_python_string
We have a function host_string_to_python_string, which is a wrapper
around PyString_Decode, which, on Python 3, is an alias for
PyUnicode_Decode.

However, there are a few places where we still call PyUnicode_Decode
directly.

This commit updates some places where we call PyUnicode_Decode to
call host_string_to_python_string instead.  I've only updated those
places where PyUnicode_Decode is being used to create a PyObject* that
is about to be returned to Python.  There are a couple of places (in
py-cmd.c) where PyUnicode_Decode is used to create an argument value
for a hook into Python code, I've not touched this case for now.

To make the reuse of host_string_to_python_string easier, I've added a
couple of overloads for this function in python-internal.h, these all
just forward their calls onto the single base implementation.

For Python 3 PyString_Decode is setup (in python-internal.h) as an
alias for PyUnicode_Decode, so there should certainly be no changes in
this case.

For Python 2 this commit will change the behaviour.  When we
previously return a result from PyUnicode_Decode we would have been
returning a Unicode object, while now we will return a str object.
I've checked the documentation, and, as far as I can tell, the methods
I've touched were all documented as returning a string, so, I would
claim, the new behaviour now better matches the documentation.
However, I understand that people might feel that returning a Unicode
object is better practice.

That said, I've been reading this page:

  https://pythonhosted.org/kitchen/unicode-frustrations.html

item #3 especially recommends that unicode strings should be converted
to str objects before being printed.  That would mean that users
should have been adding .encode() calls to the output of the routines
I've changed in this commit.

In a later commit I plan to address handling of non-ascii characters
in some additional cases (for Python 2), and this is made easier by
returning a str object.