wip: gdb: don't use PyString_Decode for Python 2
The PyString_Decode function does the following:

  PyString_FromStringAndSize (...)
  PyString_AsDecodedObject (...)
  if (PyUnicode_Check(v)) {
    v = PyUnicode_AsEncodedString(...);

My understanding of what this does is:

  1. Create an str object from a c-string,
  2. Convert the str to a Unicode using the encoding passed into
     PyString_Decode, and finally,
  3. Convert the Unicode object back to a string using the default
     encoding scheme.

The problem here is that the default encoding scheme (ascii) will not
be happy if there are any non-ascii characters in the original
c-string.  This results in things like this (when using GDB built
against Python 2):

  $ gdb -q --batch -ex 'python print gdb.Value(u"\xff\x01abc")'
  Traceback (most recent call last):
    File "<string>", line 1, in <module>
  UnicodeEncodeError: 'ascii' codec can't encode character u'\xff' in position 1: ordinal not in range(128)
  Error while executing Python code.

I wonder if we should avoid the str -> unicode -> str process for
Python 2?  This certainly fixes the above issue, and doesn't introduce
any regressions for the gdb.python/*.exp tests when using Python 2.
We now see:

  $ gdb -q --batch -ex 'python print gdb.Value(u"\xff\x01abc")'
  "ÿ\001abc"

which is the same as we see for Python 3.