[wip] attempt to fix gdb testsuite runs with multiple boards
I noticed that running the gdb testsuite with multiple boards no
longer works, for example:

  make check-gdb RUNTESTFLAGS="--target_board='native-gdbserver unix' gdb.base/break.exp"

will fail.

This is because gdb/testsuite/boards/gdbserver-base.exp sets the
GDBFLAGS variable for the first board, which is then inherited by all
subsequent boards.

My attempt to fix this was to have boards store the options not in the
GDBFLAGS variable, but in the board_info global instead, hashed by
target board name.

This works for the example above, but introduces other problems.

Board files include other board files, however, this inclusion nesting
is protected, a file can only be loaded once.  And so, consider
gdb/boards/local-board.exp, which also adds to the GDBFLAGS.

Now I have local-board.exp setting flags within board_info, however,
these changes are only done once, when local-board.exp is loaded for
the first target_board, the flags are NOT adjusted for the second
board.  The result is that the changes made by local-board.exp are
only noticed by the first target board.  We're not really any better
off than before, just broken in a different way.

I don't currently have a solution for this problem, so I'm parking
this patch for now.